fix: set app.proxy=true to websocket server
