+++ src/server/server.ts
@@ -20,6 +20,7 @@ import KoaJsonRpc from './koaJsonRpc';
 import { spec } from './koaJsonRpc/lib/RpcError';
 import { getLimitDuration } from './koaJsonRpc/lib/utils';
 import EthereumRPCConformityService from './koaJsonRpc/services/EthereumRPCConformityService';
+import { applyProxyMiddleware } from './utils/proxyUtils';
 
 // https://nodejs.org/api/async_context.html#asynchronous-context-tracking
 const context = new AsyncLocalStorage<{ requestId: string }>();
@@ -52,88 +53,6 @@ const mainLogger = pino({
 const ethereumRPCConformityService = new EthereumRPCConformityService();
 
 export const logger = mainLogger.child({ name: 'rpc-server' });
-
-/**
- * Parse RFC 7239 Forwarded header to extract the original client IP
- *
- * This function safely parses the Forwarded header without using regex to avoid
- * ReDoS (Regular Expression Denial of Service) vulnerabilities. It includes
- * input length limits and basic validation to prevent malicious input from
- * causing performance issues.
- *
- * @param forwardedHeader - The Forwarded header value
- * @returns The client IP address or null if not found
- */
-function parseForwardedHeader(forwardedHeader: string): string | null {
-  try {
-    // Limit input length to prevent DoS attacks
-    if (forwardedHeader.length > 1000) {
-      return null;
-    }
-
-    // Split by comma to handle multiple forwarded entries
-    const entries = forwardedHeader.split(',');
-
-    // Take the first entry (original client)
-    const firstEntry = entries[0]?.trim();
-    if (!firstEntry) return null;
-
-    // Find the 'for=' parameter using safe string parsing
-    const forIndex = firstEntry.toLowerCase().indexOf('for=');
-    if (forIndex === -1) return null;
-
-    // Extract the value after 'for='
-    const valueStart = forIndex + 4; // Length of 'for='
-    if (valueStart >= firstEntry.length) return null;
-
-    let ip: string;
-    const char = firstEntry[valueStart];
-
-    if (char === '"') {
-      // Quoted value: for="192.168.1.1" or for="[2001:db8::1]"
-      const closeQuoteIndex = firstEntry.indexOf('"', valueStart + 1);
-      if (closeQuoteIndex === -1) return null;
-      ip = firstEntry.substring(valueStart + 1, closeQuoteIndex);
-
-      // Handle IPv6 in brackets within quotes: for="[2001:db8::1]"
-      if (ip.startsWith('[') && ip.endsWith(']')) {
-        ip = ip.substring(1, ip.length - 1);
-      }
-    } else if (char === '[') {
-      // IPv6 in brackets: for=[2001:db8::1]
-      const closeBracketIndex = firstEntry.indexOf(']', valueStart + 1);
-      if (closeBracketIndex === -1) return null;
-      ip = firstEntry.substring(valueStart + 1, closeBracketIndex);
-    } else {
-      // Unquoted value: for=192.168.1.1
-      let endIndex = valueStart;
-      while (endIndex < firstEntry.length) {
-        const c = firstEntry[endIndex];
-        if (c === ';' || c === ',' || c === ' ' || c === '\t') {
-          break;
-        }
-        endIndex++;
-      }
-      ip = firstEntry.substring(valueStart, endIndex);
-    }
-
-    // Basic validation: ensure we have a non-empty result
-    if (!ip || ip.length === 0 || ip.length > 45) {
-      // Max IPv6 length is 45 chars
-      return null;
-    }
-
-    // Basic IP format validation (very permissive)
-    if (!/^[a-fA-F0-9:.]+$/.test(ip)) {
-      return null;
-    }
-
-    return ip;
-  } catch {
-    // If parsing fails, return null to avoid breaking the request
-    return null;
-  }
-}
 
 /**
  * Initialize the server components
@@ -183,27 +102,8 @@ export async function initializeServer(sharedRelay?: Relay, sharedRegister?: Reg
     buckets: [5, 10, 25, 50, 100, 250, 500, 1000, 2500, 5000, 10000, 20000, 30000, 40000, 50000, 60000], // ms (milliseconds)
   });
 
-  // enable proxy support to trust proxy-added headers for client IP detection
-  app.proxy = true;
-
-  // Middleware to parse RFC 7239 Forwarded header and make it compatible with Koa's X-Forwarded-For parsing
-  app.use(async (ctx, next) => {
-    // Only process if X-Forwarded-For doesn't exist but Forwarded does
-    if (!ctx.request.headers['x-forwarded-for'] && ctx.request.headers['forwarded']) {
-      const forwardedHeader = ctx.request.headers['forwarded'] as string;
-
-      // Parse the Forwarded header to extract the client IP
-      // Format: Forwarded: for="192.168.1.1";by="10.0.0.1", for="203.0.113.1";by="10.0.0.2"
-      const clientIp = parseForwardedHeader(forwardedHeader);
-
-      if (clientIp) {
-        // Set X-Forwarded-For so Koa can parse it normally
-        ctx.request.headers['x-forwarded-for'] = clientIp;
-      }
-    }
-
-    await next();
-  });
+  // Enable proxy support and RFC 7239 Forwarded header translation
+  applyProxyMiddleware(app);
 
   // Set CORS
   app.use(cors({ allowMethods: ['GET', 'POST'] }));
